requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==4.1.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
commonmark==0.9.1
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
Cython==3.0.10
defusedxml==0.7.1
deptry==0.16.1
doxapy==0.9.2
dtaidistance==2.3.11
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
grpcio-tools==1.62.1
h5py==3.11.0
idna==3.7
imageio==2.34.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.2.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lxml==5.2.1
manimce==0.1.0
manimlib==0.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pangocairocffi==0.3.2
pangocffi==0.6.0
pep517==0.13.1
pillow==10.3.0
progressbar==2.5
protobuf==4.25.3
pycairo==1.26.0
pyclipper==1.3.0.post5
pycparser==2.22
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.3
six==1.16.0
svglib==1.5.1
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
